translate-and-analysic/requirements.txt
azure-cognitiveservices-vision-computervision
azure-ai-translation-text==1.0.0b1
flask
azure-storage-blob
azure-identity

azure-ai-textanalytics==5.2.0
